Moved comment handling out of CoNLL2RDF
